feat(agents-tools): split tools into cohesive modules (router_tool, planning, retrieval, synthesis, validation) and add telemetry; remove monolithic tools.py
