+++ src/agents/tools/constants.py
@@ -1,0 +1,37 @@
+"""Shared constants for agent tools modules."""
+
+# Query routing thresholds
+COMPLEX_QUERY_WORD_THRESHOLD = 20
+MEDIUM_QUERY_WORD_THRESHOLD = 10
+RECENT_CHAT_HISTORY_LIMIT = 3
+
+# Confidence defaults
+COMPLEX_CONFIDENCE = 0.9
+MEDIUM_CONFIDENCE = 0.85
+SIMPLE_CONFIDENCE = 0.95
+CONFIDENCE_ADJUSTMENT_FACTOR = 0.8
+FALLBACK_CONFIDENCE = 0.5
+
+# Retrieval tuning
+VARIANT_QUERY_LIMIT = 2
+CONTENT_KEY_LENGTH = 100
+MAX_RETRIEVAL_RESULTS = 10
+SIMILARITY_THRESHOLD = 0.8
+
+# Validation thresholds
+MIN_RESPONSE_LENGTH = 50
+CONFIDENCE_REDUCTION_INCOMPLETE = 0.8
+CONFIDENCE_REDUCTION_NO_SOURCE = 0.7
+CONFIDENCE_REDUCTION_NO_SOURCES = 0.6
+CONFIDENCE_REDUCTION_HALLUCINATION = 0.5
+CONFIDENCE_REDUCTION_LOW_RELEVANCE = 0.8
+CONFIDENCE_REDUCTION_COHERENCE = 0.9
+RELEVANCE_THRESHOLD = 0.3
+ACCEPT_CONFIDENCE_THRESHOLD = 0.8
+REFINE_CONFIDENCE_THRESHOLD = 0.6
+VALIDATION_CONFIDENCE_THRESHOLD = 0.6
+MAX_SENTENCE_WORD_OVERLAP = 5
+MIN_AVG_SENTENCE_LENGTH = 3
+MAX_AVG_SENTENCE_LENGTH = 50
+MAX_ISSUES_FOR_ACCEPT = 1
+FIRST_N_SOURCES_CHECK = 3
